move withDelayedProgress to its own file in the ui folder
