+++ src/commands.ts
@@ -21,7 +21,8 @@ import { createSwiftTask, SwiftTaskProvider } from "./SwiftTaskProvider";
 import { FolderContext } from "./FolderContext";
 import { PackageNode } from "./ui/PackageDependencyProvider";
 import { withQuickPick } from "./ui/QuickPick";
-import { execSwift, getErrorDescription, withDelayedProgress } from "./utilities/utilities";
+import { withDelayedProgress } from "./ui/withDelayedProgress";
+import { execSwift, getErrorDescription } from "./utilities/utilities";
 import { Version } from "./utilities/version";
 import { DarwinCompatibleTarget, SwiftToolchain } from "./toolchain/toolchain";
 import { debugSnippet, runSnippet } from "./SwiftSnippets";
